+++ plane/src/controller/terminate.rs
@@ -2,7 +2,7 @@ use super::{core::Controller, error::IntoApiError};
 use crate::{
     names::BackendName,
     protocol::BackendAction,
-    types::{ClusterName, TerminationKind},
+    types::{ClusterName, TerminationKind, backend_state::TerminationReason},
 };
 use axum::{
     extract::{Path, State},
@@ -35,7 +35,7 @@ async fn terminate(
         .create_pending_action(
             backend_id,
             backend.drone_id,
-            &BackendAction::Terminate { kind },
+            &BackendAction::Terminate { kind, reason: TerminationReason::External },
         )
         .await
         .or_internal_error("Database error")?;
